requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
aiosqlite==0.19.0
alembic==1.13.1
asyncpg==0.29.0
cachetools==5.3.2
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.27
greenlet==3.0.3
geopy==2.4.1